fix: change sanitizer flags from INTERFACE to PUBLIC
INTERFACE visibility only propagated sanitizer flags to consumers of
TinyKafka_core (kafka binary, test executables), but the static library's
own translation units were not instrumented for ASan/UBSan detection.

Also fix -Wconversion warnings in endian.hpp and integration_test.cpp
exposed by UBSan's implicit-conversion checker.

+++ src/util/endian.hpp
@@ -27,7 +27,7 @@ constexpr void write_int64_be(int64_t value, std::span<std::uint8_t, 8> out) noe
 }
 
 constexpr auto decode_int16_be(std::span<const std::uint8_t, 2> data) noexcept -> int16_t {
-    return (static_cast<int16_t>(data[0]) << 8) | static_cast<int16_t>(data[1]);
+    return static_cast<int16_t>((static_cast<int16_t>(data[0]) << 8) | static_cast<int16_t>(data[1]));
 }
 
 constexpr auto decode_int32_be(std::span<const std::uint8_t, 4> data) noexcept -> int32_t {
